chore: remove dashboardItem interface to models folder
